Added functionality for users to join existing contests
also cleaned up code and updated readme
