+++ src/main/java/com/hatstone/bobertapi/dto/Submission.java
@@ -3,28 +3,28 @@ package com.hatstone.bobertapi.dto;
 import java.io.Serializable;
 
 public class Submission {
-    private Long userid;
-    private Long problemid;
+    private Long userId;
+    private Long problemId;
     private byte[] data;
     private String language;
 
     // Constructors //
-    public Submission(Long userid, Long problemid, byte[] data, String language){
-        this.userid = userid;
-        this.problemid = problemid;
+    public Submission(Long userId, Long problemId, byte[] data, String language){
+        this.userId = userId;
+        this.problemId = problemId;
         this.data = data;
         this.language = language;
     }
 
     // Getters //
-    public Long getUserId() {return userid;}
-    public Long getProblemId() {return problemid;}
+    public Long getUserId() {return userId;}
+    public Long getProblemId() {return problemId;}
     public byte[] getData() {return data;}
     public String getLanguage() {return language;}
 
     // Setters //
-    public void setUserId(Long userid) {this.userid = userid;}
-    public void setProblemId(Long problemid) {this.problemid = problemid;}
+    public void setUserId(Long userId) {this.userId = userId;}
+    public void setProblemId(Long problemId) {this.problemId = problemId;}
     public void setData(byte[] data) {this.data = data;}
     public void setLanguage(String language) {this.language = language;}
 }
